refactor(qa): extract native vitest report validation
